Upgrade Job-DSL dependency to 1.42
Replace job with freeStyleJob and view with listView in most tests.

Remove tests that check for an exception to be thrown when no job name is
provided, since the new Job DSL syntax requires the job name as a
parameter. Also the name method does not exist in Job DSL anymore, so
replace usages with calls to the setter method for name.

javaposse.jobdsl.dsl.JobType was removed, so instead use the simlar class
com.terrafolio.gradle.plugins.jenkins.jobdsl.DSLJobType.

Also add a direct dependency to Guava because it was removed from Job DSL.